+++ src/pages/main/header/HeaderLogo.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import "../../../styles/common.css";
+import "./index.css";
+
+const HeaderLogo = () => {
+  return <div className="logo flex justify-center align-center"></div>;
+};
+
+export default HeaderLogo;
