+++ src/pages/main/header/index.css
@@ -5,9 +5,9 @@
 }
 
 .logo {
-  height: 100%;
+  height: 1.5rem;
   width: 72px;
-  background: center no-repeat url("../assets/img/logo.png");
+  background: center no-repeat url("../../../assets/img/logo.png");
 }
 
 .nav-wrap {
