+++ src/pages/main/header/index.tsx
@@ -1,75 +1,18 @@
 import React, { useState } from "react";
-import "../../styles/common.css";
-import "./index.css"
-import { BsSun, BsMoon, BsSearch } from "react-icons/bs";
-import { GoTriangleDown } from "react-icons/go";
-import ProfileImage from "../../../components/ProfileImage";
+import NavMenu from "../../../components/ui/NavMenu";
+import "../../../styles/common.css";
+import HeaderLogo from "./HeaderLogo";
+import "./index.css";
 interface MainHeaderProps {
   darkMode: boolean;
   setDarkMode: any;
 }
 
 const MainHeader: React.FC<MainHeaderProps> = ({ darkMode, setDarkMode }) => {
-  const [toggleBar, setToggleBar] = useState<boolean>(false);
-
-  const toggleMode = (state: boolean, setFunc: any) => setFunc(!state);
-
   return (
-    <div className=" flex justify-center align-center main-background-color">
-      <div className="MainHeader-wrap flex space-between align-center">
-        <div className="logo flex justify-center align-center"></div>
-        
-        <div className="nav-wrap flex">
-          {darkMode ? (
-            <div
-              className="nav-icon flex justify-center align-center"
-              onClick={() => toggleMode(darkMode, setDarkMode)}
-            >
-              <BsSun size={30} />
-            </div>
-          ) : (
-            <div
-              className="nav-icon flex justify-center align-center"
-              onClick={() => toggleMode(darkMode, setDarkMode)}
-            >
-              <BsMoon size={30} />
-            </div>
-          )}
-
-          <div className="nav-icon flex justify-center align-center">
-            <BsSearch size={25} />
-          </div>
-          <div className="new-article flex justify-center align-center">
-            새글 작성
-          </div>
-          <div className="profile-wrap flex space-between align-center">
-            <ProfileImage
-              width={50}
-              height={50}
-              url={
-                "https://velog.velcdn.com/images/ninto_2/post/e4eaa7ec-223f-4ada-be19-11dadf1d8d0f/image.gif"
-              }
-            />
-            <div
-              className="toggle_btn flex justify-center align-center"
-              onClick={() => toggleMode(toggleBar, setToggleBar)}
-            >
-              <GoTriangleDown />
-            </div>
-          </div>
-        </div>
-        {toggleBar && (
-          <div className="bars flex flex-column">
-            <ul>
-              <li>내 벨로그</li>
-              <li>임시 글</li>
-              <li>읽기 목록</li>
-              <li>설정</li>
-              <li>로그아웃</li>
-            </ul>
-          </div>
-        )}
-      </div>
+    <div className=" flex space-between align-center main-background-color">
+      <HeaderLogo/>
+      <NavMenu darkMode={darkMode} setDarkMode={setDarkMode} />
     </div>
   );
 };
